# docs/api/openapi-b2b-workflow.yaml
openapi: 3.0.3
info:
  title: EPACKAGE Lab B2B Workflow API
  description: |
    10단계 B2B 주문 프로세스를 위한 RESTful API

    워크플로우:
    1. Registration (회원가입)
    2. Quotation (견적)
    3. Order (주문)
    4. Data Entry (데이터입고)
    5. Work Order (작업표준서)
    6. Contract (계약서)
    7. Signing (전자서명)
    8. Production (생산)
    9. Stock In (입고)
    10. Shipment (출하)
  version: 1.0.0
  contact:
    name: EPACKAGE Lab
    email: api@epackage-lab.com
servers:
  - url: https://api.epackage-lab.com/api/v1
    description: Production server
  - url: https://staging-api.epackage-lab.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /orders/{id}:
    get:
      tags: [Orders]
      summary: 주문 상세 조회
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 조회 성공
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: 주문 상태 변경
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequest'
      responses:
        '200':
          description: 상태 변경 완료
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 토큰 인증
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    UpdateOrderStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
        reason:
          type: string
        metadata:
          type: object
...